improve: custom views
This commit improve custom views by moving them to a separated file, reducing project size.
