+++ TODO.md
@@ -3,7 +3,6 @@
 - (Optimization) Change buttons of the same page for texts.
 - (Optimization) Option to hash the name of the achievements.
 - (Optimization) Option to make a switch with data of all scenes to avoid need of more JSON reads.
-- (Optimization) Move custom functions to a custom.js file.
 - (Bug fix) Sub scene parent to the sub scene activates speech animation.
 - (Customizability) Allow to remove footer and add buttons manually.
 - (Code quality) Only add necessary `import`s.
