+++ index.js
@@ -3,11 +3,13 @@ import scene from './src/scene.js'
 import subScene from './src/sub-scene.js'
 import menu from './src/menu.js'
 import achievements from './src/achievements.js'
+import custom from './src/custom.js'
 
 export default {
   coder,
   scene,
   subScene,
   menu,
-  achievements
+  achievements,
+  custom
 }
